fix: address PR review feedback for /models endpoint
- Handle double-prefixing when canonical != original provider name (e.g. zhipuai -> zai)
- Add 5s per-provider timeout to prevent endpoint hanging on slow providers
- Pass real Unix timestamp for 'created' field instead of hardcoding 0
- Add defensive finite/non-negative check for pricing values
- Return empty string for unknown pricing instead of misleading "0"
- Add #[serde(skip_serializing_if)] on optional fields to avoid null serialization
- Split types.rs into sub-files to respect 50-line file limit

CodeTether-Provenance-ID: ctprov_7a28296b3de3416f902d14a28a2d4a94
CodeTether-Origin: local_cli
CodeTether-Agent-Name: build
CodeTether-Session-ID: ebb4d373-2de8-4e4e-9223-0d7cd2b01c48
