+++ src/server/models_catalog/pricing.rs
@@ -9,8 +9,8 @@ pub(crate) fn price(model: &crate::provider::ModelInfo) -> super::types::Pricing
 
 fn per_token(cost_per_million: Option<f64>) -> String {
     match cost_per_million {
-        Some(cost) => trim_float(cost / 1_000_000.0),
-        None => "0".to_string(),
+        Some(cost) if cost.is_finite() && cost >= 0.0 => trim_float(cost / 1_000_000.0),
+        _ => "".to_string(),
     }
 }
