+++ src/server/models_catalog/types_pricing.rs
@@ -1,0 +1,9 @@
+//! Pricing response struct.
+
+use serde::Serialize;
+
+#[derive(Debug, Serialize)]
+pub(crate) struct Pricing {
+    pub(crate) prompt: String,
+    pub(crate) completion: String,
+}
